+++ src/transceivercontroller/DefaultTransceiverController.ts
@@ -548,7 +548,8 @@ export default class DefaultTransceiverController
         this.audioRedWorker,
         { type: 'ReceiverTransform' }
       );
-    } /* istanbul ignore else */ else if (supportsInsertableStreams) {
+      // eslint-disable-next-line
+    } else /* istanbul ignore else */ if (supportsInsertableStreams) {
       // @ts-ignore
       const sendStreams = this._localAudioTransceiver.sender.createEncodedStreams();
       // @ts-ignore
@@ -602,7 +603,8 @@ export default class DefaultTransceiverController
       transceiver.receiver.transform = new RTCRtpScriptTransform(this.audioRedWorker, {
         type: 'PassthroughTransform',
       });
-    } /* istanbul ignore else */ else if (supportsInsertableStreams) {
+      // eslint-disable-next-line
+    } else /* istanbul ignore else */ if (supportsInsertableStreams) {
       // @ts-ignore
       const sendStreams = transceiver.sender.createEncodedStreams();
       // @ts-ignore