+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for node 20 and drop support for node < 18.
+- Add support for H.264 profiles besides Constrained Baseline Profile.
 
 ### Removed